@@ -2,7 +2,7 @@
     "name": "fsi/resource-repository-bundle",
     "type": "symfony-bundle",
     "description": "Resource repository bundle",
-    "keywords": ["fsi", "resource", "repository", "bundle"],
+    "keywords": ["fsi", "resource", "repository", "bundle", "symfony"],
     "license": "MIT",
     "authors": [
         {
@@ -11,23 +11,26 @@
         }
     ],
     "require": {
-        "php": ">=5.4",
-        "twig/twig": "~1.13",
-        "symfony/framework-bundle": "^2.2|^3.0",
-        "symfony/form" : "^2.2|^3.0",
-        "symfony/validator": "^2.2|^3.0",
-        "symfony/event-dispatcher": "^2.2|^3.0",
+        "php": ">=7.1",
+        "twig/twig": "^1.34|^2.0",
+        "doctrine/dbal": "^2.6",
         "doctrine/doctrine-bundle": "^1.4",
-        "symfony/doctrine-bridge": "^2.3|^3.0",
-        "symfony/dependency-injection": "^2.6|^3.0",
-        "symfony/property-access": "^2.3|^3.0"
+        "symfony/framework-bundle": "^3.0|^4.0",
+        "symfony/form" : "^3.0|^4.0",
+        "symfony/validator": "^3.0|^4.0",
+        "symfony/event-dispatcher": "^3.0|^4.0",
+        "symfony/doctrine-bridge": "^3.0|^4.0",
+        "symfony/dependency-injection": "^3.0|^4.0",
+        "symfony/property-access": "^3.0|^4.0"
     },
     "require-dev": {
-        "doctrine/orm": "~2.3",
-        "friendsofsymfony/ckeditor-bundle": "dev-master",
-        "phpspec/phpspec": "^2.0",
-        "fsi/doctrine-extensions-bundle": "~1.0",
-        "fsi/form-extensions-bundle": "1.0.*"
+        "doctrine/orm": "^2.5",
+        "friendsofsymfony/ckeditor-bundle": "^1.0",
+        "phpspec/phpspec": "^4.0",
+        "phpspec/prophecy": "^1.7",
+        "fsi/doctrine-extensions": "^1.0|^2.0@dev",
+        "fsi/doctrine-extensions-bundle": "^1.0|^2.0@dev",
+        "fsi/form-extensions-bundle": "^2.0@dev"
     },
     "suggest": {
         "fsi/doctrine-extensions-bundle": "For using fsi_file resource type",
@@ -41,7 +44,8 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "1.1-dev"
+            "1.1": "1.1-dev",
+            "dev-master": "2.0-dev"
         }
     }
 }
