@@ -7,20 +7,19 @@
  * file that was distributed with this source code.
  */
 
+declare(strict_types=1);
+
 namespace FSi\Bundle\ResourceRepositoryBundle\DependencyInjection\Compiler;
 
 use FSi\Bundle\ResourceRepositoryBundle\Exception\CompilerPassException;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 
-/**
- * @author Norbert Orzechowicz <norbert@fsi.pl>
- */
 class ResourcePass implements CompilerPassInterface
 {
     public function process(ContainerBuilder $container)
     {
-        $resources = array();
+        $resources = [];
         foreach ($container->findTaggedServiceIds('resource.type') as $serviceId => $tag) {
             if (!isset($tag[0]['alias'])) {
                 throw new CompilerPassException(sprintf('Service %s missing alias attribute', $serviceId));
