@@ -7,8 +7,11 @@
  * file that was distributed with this source code.
  */
 
+declare(strict_types=1);
+
 namespace FSi\Bundle\ResourceRepositoryBundle\DependencyInjection\Compiler;
 
+use FSi\Bundle\ResourceRepositoryBundle\Repository\Resource\Type\CKEditorType;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\Definition;
@@ -17,8 +20,8 @@ class ResourceCKEditorPass implements CompilerPassInterface
 {
     public function process(ContainerBuilder $container)
     {
-        $definition = new Definition('FSi\Bundle\ResourceRepositoryBundle\Repository\Resource\Type\CKEditorType');
-        $definition->addTag('resource.type', array('alias' => 'ckeditor'));
+        $definition = new Definition(CKEditorType::class);
+        $definition->addTag('resource.type', ['alias' => 'ckeditor']);
 
         $container->setDefinition('fsi_resource_repository.resource.type.ckeditor', $definition);
     }
