@@ -7,6 +7,8 @@
  * file that was distributed with this source code.
  */
 
+declare(strict_types=1);
+
 namespace FSi\Bundle\ResourceRepositoryBundle\Form\Type;
 
 use FSi\Bundle\ResourceRepositoryBundle\Exception\ResourceFormTypeException;
@@ -14,7 +16,6 @@ use FSi\Bundle\ResourceRepositoryBundle\Repository\MapBuilder;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
-use Symfony\Component\OptionsResolver\OptionsResolverInterface;
 
 class ResourceType extends AbstractType
 {
@@ -28,53 +29,24 @@ class ResourceType extends AbstractType
      */
     protected $resourceClass;
 
-    /**
-     * @param MapBuilder $mapBuilder
-     * @param $resourceClass
-     */
-    public function __construct(MapBuilder $mapBuilder, $resourceClass)
+    public function __construct(MapBuilder $mapBuilder, string $resourceClass)
     {
         $this->mapBuilder = $mapBuilder;
         $this->resourceClass = $resourceClass;
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function getName()
     {
         return 'resource';
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    public function setDefaultOptions(OptionsResolverInterface $resolver)
-    {
-        $this->configureOptions($resolver);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
     public function configureOptions(OptionsResolver $resolver)
     {
-        $resolver->setDefaults(
-            array(
-                'data_class' => $this->resourceClass
-            )
-        );
+        $resolver->setDefaults(['data_class' => $this->resourceClass]);
 
-        $resolver->setRequired(
-            array(
-                'resource_key'
-            )
-        );
+        $resolver->setRequired(['resource_key']);
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
         if (!$this->mapBuilder->hasResource($options['resource_key'])) {
