@@ -7,23 +7,21 @@
  * file that was distributed with this source code.
  */
 
+declare(strict_types=1);
+
 namespace FSi\Bundle\ResourceRepositoryBundle\Repository\Resource\Type;
+
+use FSi\Bundle\DoctrineExtensionsBundle\Form\Type\FSi\FileType as FSiFileFormType;
 
 class FSiFileType extends AbstractType
 {
-    /**
-     * {@inheritdoc}
-     */
-    public function getResourceProperty()
+    public function getResourceProperty(): string
     {
         return 'fileValue';
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    protected function getFormType()
+    protected function getFormType(): string
     {
-        return 'fsi_file';
+        return FSiFileFormType::class;
     }
 }
