@@ -7,34 +7,32 @@
  * file that was distributed with this source code.
  */
 
+declare(strict_types=1);
+
 namespace FSi\Bundle\ResourceRepositoryBundle\Repository\Resource\Type;
+
+use FSi\Bundle\DoctrineExtensionsBundle\Form\Type\FSi\RemovableFileType as FSiRemovableFileFormType;
 
 class FSiRemovableFileType extends AbstractType
 {
-    /**
-     * {@inheritdoc}
-     */
-    public function getResourceProperty()
+    public function getResourceProperty(): string
     {
         return 'fileValue';
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    protected function getFormType()
+    protected function getFormType(): string
     {
-        return 'fsi_removable_file';
+        return FSiRemovableFileFormType::class;
     }
 
-    protected function buildFormOptions()
+    protected function buildFormOptions(): array
     {
         $options = parent::buildFormOptions();
 
         if (isset($options['constraints'])) {
             $options['file_options'] = array_merge(
-                isset($options['file_options']) ? $options['file_options'] : array(),
-                array('constraints' => $options['constraints'])
+                isset($options['file_options']) ? $options['file_options'] : [],
+                ['constraints' => $options['constraints']]
             );
             unset($options['constraints']);
         }
