@@ -1,22 +1,29 @@
 <?php
 
+/**
+ * (c) FSi sp. z o.o. <info@fsi.pl>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
 namespace spec\FSi\Bundle\ResourceRepositoryBundle\DependencyInjection;
 
+use FSi\Bundle\ResourceRepositoryBundle\DependencyInjection\Configuration;
 use PhpSpec\ObjectBehavior;
-use Prophecy\Argument;
+use Symfony\Component\Config\Definition\Builder\TreeBuilder;
 
-/**
- * @author Norbert Orzechowicz <norbert@fsi.pl>
- */
 class ConfigurationSpec extends ObjectBehavior
 {
     function it_is_initializable()
     {
-        $this->shouldHaveType('\FSi\Bundle\ResourceRepositoryBundle\DependencyInjection\Configuration');
+        $this->shouldHaveType(Configuration::class);
     }
 
     function it_should_return_tree()
     {
-        $this->getConfigTreeBuilder()->shouldReturnAnInstanceOf('\Symfony\Component\Config\Definition\Builder\TreeBuilder');
+        $this->getConfigTreeBuilder()->shouldReturnAnInstanceOf(TreeBuilder::class);
     }
 }
