@@ -1,14 +1,26 @@
 <?php
 
+/**
+ * (c) FSi sp. z o.o. <info@fsi.pl>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
 namespace spec\FSi\Bundle\ResourceRepositoryBundle\Doctrine;
 
 use Doctrine\ORM\EntityManager;
+use Doctrine\ORM\EntityRepository;
 use Doctrine\ORM\Mapping\ClassMetadata;
-use FSi\Bundle\ResourceRepositoryBundle\Model\Resource as BaseResource;
+use FSi\Bundle\ResourceRepositoryBundle\Doctrine\ResourceRepository;
 use FSi\Bundle\ResourceRepositoryBundle\Exception\EntityRepositoryException;
+use FSi\Bundle\ResourceRepositoryBundle\Model\Resource as BaseResource;
 use FSi\Bundle\ResourceRepositoryBundle\Model\ResourceValue;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
+use spec\FSi\Bundle\ResourceRepositoryBundle\Doctrine\Resource;
 
 class Resource extends BaseResource
 {
@@ -18,39 +30,39 @@ class ResourceRepositorySpec extends ObjectBehavior
 {
     function let(EntityManager $em, ClassMetadata $class)
     {
-        $class->name = 'spec\\FSi\\Bundle\\ResourceRepositoryBundle\\Doctrine\\Resource';
+        $class->name = Resource::class;
         $this->beConstructedWith($em, $class);
     }
 
     function it_is_initializable()
     {
-        $this->shouldHaveType('FSi\Bundle\ResourceRepositoryBundle\Doctrine\ResourceRepository');
+        $this->shouldHaveType(ResourceRepository::class);
     }
 
     function it_is_doctrine_entity_repository()
     {
-        $this->shouldBeAnInstanceOf('Doctrine\ORM\EntityRepository');
+        $this->shouldBeAnInstanceOf(EntityRepository::class);
     }
 
     function it_throw_exception_during_findBy_method()
     {
         $this->shouldThrow(
             new EntityRepositoryException('Method "findBy" is not supported in "FSiResourceRepository:Resource" entity repository')
-        )->during('findBy', array(array()));
+        )->during('findBy', [[]]);
     }
 
     function it_throw_exception_during_findAll_method()
     {
         $this->shouldThrow(
             new EntityRepositoryException('Method "findAll" is not supported in "FSiResourceRepository:Resource" entity repository')
-        )->during('findAll', array());
+        )->during('findAll', []);
     }
 
     function it_throw_exception_during_findOneBy_method()
     {
         $this->shouldThrow(
             new EntityRepositoryException('Method "findOneBy" is not supported in "FSiResourceRepository:Resource" entity repository')
-        )->during('findOneBy', array(array()));
+        )->during('findOneBy', [[]]);
     }
 
     function it_return_entity_even_if_it_not_exist_in_db(EntityManager $em)
@@ -91,7 +103,7 @@ class ResourceRepositorySpec extends ObjectBehavior
 
     public function getMatchers()
     {
-        return array(
+        return [
             'returnResourceWithKey' => function($subject, $key) {
                 if (!$subject instanceof Resource) {
                     return false;
@@ -103,6 +115,6 @@ class ResourceRepositorySpec extends ObjectBehavior
 
                 return true;
             }
-        );
+        ];
     }
 }
