@@ -1,13 +1,23 @@
 <?php
 
+/**
+ * (c) FSi sp. z o.o. <info@fsi.pl>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
 namespace spec\FSi\Bundle\ResourceRepositoryBundle\Form\Type;
 
-use FSi\Bundle\ResourceRepositoryBundle\Model\Resource;
+use FSi\Bundle\DemoBundle\Entity\Resource;
 use FSi\Bundle\ResourceRepositoryBundle\Exception\ResourceFormTypeException;
+use FSi\Bundle\ResourceRepositoryBundle\Form\Type\ResourceType;
 use FSi\Bundle\ResourceRepositoryBundle\Repository\MapBuilder;
 use FSi\Bundle\ResourceRepositoryBundle\Repository\Resource\Type\TextType;
 use PhpSpec\ObjectBehavior;
-use Prophecy\Argument;
+use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilder;
 use Symfony\Component\Form\FormFactory;
 use Symfony\Component\OptionsResolver\OptionsResolver;
@@ -16,17 +26,17 @@ class ResourceTypeSpec extends ObjectBehavior
 {
     function let(MapBuilder $map)
     {
-        $this->beConstructedWith($map, 'FSi\Bundle\DemoBundle\Entity\Resource');
+        $this->beConstructedWith($map, Resource::class);
     }
 
     function it_is_initializable()
     {
-        $this->shouldHaveType('FSi\Bundle\ResourceRepositoryBundle\Form\Type\ResourceType');
+        $this->shouldHaveType(ResourceType::class);
     }
 
     function it_is_form_type()
     {
-        $this->shouldBeAnInstanceOf('Symfony\Component\Form\AbstractType');
+        $this->shouldBeAnInstanceOf(AbstractType::class);
     }
 
     function it_is_called_resource()
@@ -34,29 +44,10 @@ class ResourceTypeSpec extends ObjectBehavior
         $this->getName()->shouldReturn('resource');
     }
 
-    function it_should_have_default_data_class_and_resource_key_option(OptionsResolver $resolver)
-    {
-        $resolver->setDefaults(array(
-            'data_class' => 'FSi\Bundle\DemoBundle\Entity\Resource'
-        ))->shouldBeCalled();
-
-        $resolver->setRequired(array(
-            'resource_key'
-        ))->shouldBeCalled();
-
-        $this->setDefaultOptions($resolver);
-    }
-
     function it_configures_options(OptionsResolver $resolver)
     {
-        $resolver->setDefaults(array(
-            'data_class' => 'FSi\Bundle\DemoBundle\Entity\Resource'
-        ))->shouldBeCalled();
-
-        $resolver->setRequired(array(
-            'resource_key'
-        ))->shouldBeCalled();
-
+        $resolver->setDefaults(['data_class' => Resource::class])->shouldBeCalled();
+        $resolver->setRequired(['resource_key'])->shouldBeCalled();
         $this->configureOptions($resolver);
     }
 
@@ -66,24 +57,21 @@ class ResourceTypeSpec extends ObjectBehavior
 
         $this->shouldThrow(
             new ResourceFormTypeException('"resources.invalid_resource" is not a valid resource key')
-        )->duringBuildForm($builder, array(
-            'resource_key' => 'resources.invalid_resource'
-        ));
+        )->duringBuildForm($builder, ['resource_key' => 'resources.invalid_resource']);
     }
 
-    function it_add_form_builder_specified_by_resource_definition(MapBuilder $map, FormBuilder $builder, TextType $resource,
-            FormFactory $factory, FormBuilder $textBuilder)
-    {
+    function it_add_form_builder_specified_by_resource_definition(
+        MapBuilder $map,
+        FormBuilder $builder,
+        TextType $resource,
+        FormFactory $factory,
+        FormBuilder $textBuilder
+    ) {
         $map->hasResource('resources.resource_text')->willReturn(true);
         $map->getResource('resources.resource_text')->shouldBeCalled()->willReturn($resource);
         $builder->getFormFactory()->willReturn($factory);
         $resource->getFormBuilder($factory)->shouldBeCalled()->willReturn($textBuilder);
-
         $builder->add($textBuilder)->shouldBeCalled();
-
-
-        $this->buildForm($builder, array(
-            'resource_key' => 'resources.resource_text'
-        ));
+        $this->buildForm($builder, ['resource_key' => 'resources.resource_text']);
     }
 }
