@@ -1,9 +1,20 @@
 <?php
 
+/**
+ * (c) FSi sp. z o.o. <info@fsi.pl>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
 namespace spec\FSi\Bundle\ResourceRepositoryBundle\Repository\Resource\Type;
 
+use FSi\Bundle\ResourceRepositoryBundle\Repository\Resource\Type\CKEditorType as FSiCKEditorType;
+use FSi\Bundle\ResourceRepositoryBundle\Repository\Resource\Type\ResourceInterface;
+use FOS\CKEditorBundle\Form\Type\CKEditorType;
 use PhpSpec\ObjectBehavior;
-use Prophecy\Argument;
 use Symfony\Component\Form\FormBuilder;
 use Symfony\Component\Form\FormFactory;
 use Symfony\Component\Validator\Constraints\NotBlank;
@@ -17,12 +28,12 @@ class CKEditorTypeSpec extends ObjectBehavior
 
     function it_is_initializable()
     {
-        $this->shouldHaveType('FSi\Bundle\ResourceRepositoryBundle\Repository\Resource\Type\CKEditorType');
+        $this->shouldHaveType(FSiCKEditorType::class);
     }
 
     function it_is_resource()
     {
-        $this->shouldImplement('FSi\Bundle\ResourceRepositoryBundle\Repository\Resource\Type\ResourceInterface');
+        $this->shouldImplement(ResourceInterface::class);
     }
 
     function it_return_text_entity_field()
@@ -37,46 +48,43 @@ class CKEditorTypeSpec extends ObjectBehavior
 
     function it_return_form_builder(FormFactory $factory, FormBuilder $form)
     {
-        $factory->createNamedBuilder('textValue', $this->getCKEditorFormType(), null, array(
+        $factory->createNamedBuilder('textValue', CKEditorType::class, null, [
             'label' => false,
             'required' => false,
-        ))->shouldBeCalled()->willReturn($form);
+        ])->shouldBeCalled()->willReturn($form);
 
-        $this->getFormBuilder($factory)->shouldReturnAnInstanceOf('Symfony\Component\Form\FormBuilder');
+        $this->getFormBuilder($factory)->shouldReturnAnInstanceOf(FormBuilder::class);
     }
 
-    function it_return_form_builder_with_validation_constraints(FormFactory $factory, FormBuilder $form, NotBlank $constraint)
-    {
+    function it_return_form_builder_with_validation_constraints(
+        FormFactory $factory,
+        FormBuilder $form,
+        NotBlank $constraint
+    ) {
         $this->addConstraint($constraint);
 
-        $factory->createNamedBuilder('textValue', $this->getCKEditorFormType(), null, array(
+        $factory->createNamedBuilder('textValue', CKEditorType::class, null, [
             'label' => false,
             'required' => false,
-            'constraints' => array(
+            'constraints' => [
                 $constraint
-            )
-        ))->shouldBeCalled()->willReturn($form);
+            ]
+        ])->shouldBeCalled()->willReturn($form);
 
-        $this->getFormBuilder($factory)->shouldReturnAnInstanceOf('Symfony\Component\Form\FormBuilder');
+        $this->getFormBuilder($factory)->shouldReturnAnInstanceOf(FormBuilder::class);
     }
 
     function it_return_form_builder_with_form_options_added_to_resource_definition(FormFactory $factory, FormBuilder $form)
     {
-        $this->setFormOptions(array(
-            'attr' => array(
-                'class' => 'class-name'
-            )
-        ));
+        $this->setFormOptions(['attr' => ['class' => 'class-name']]);
 
-        $factory->createNamedBuilder('textValue', $this->getCKEditorFormType(), null, array(
+        $factory->createNamedBuilder('textValue', CKEditorType::class, null, [
             'label' => false,
             'required' => false,
-            'attr' => array(
-                'class' => 'class-name'
-            )
-        ))->shouldBeCalled()->willReturn($form);
+            'attr' => ['class' => 'class-name']
+        ])->shouldBeCalled()->willReturn($form);
 
-        $this->getFormBuilder($factory)->shouldReturnAnInstanceOf('Symfony\Component\Form\FormBuilder');
+        $this->getFormBuilder($factory)->shouldReturnAnInstanceOf(FormBuilder::class);
     }
 
     /**
