@@ -1,9 +1,19 @@
 <?php
 
+/**
+ * (c) FSi sp. z o.o. <info@fsi.pl>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
 namespace spec\FSi\Bundle\ResourceRepositoryBundle\Repository\Resource\Type;
 
+use FSi\Bundle\DoctrineExtensionsBundle\Form\Type\FSi\RemovableFileType;
+use FSi\Bundle\ResourceRepositoryBundle\Repository\Resource\Type\ResourceInterface;
 use PhpSpec\ObjectBehavior;
-use Prophecy\Argument;
 use Symfony\Component\Form\FormBuilder;
 use Symfony\Component\Form\FormFactory;
 use Symfony\Component\Validator\Constraints\NotBlank;
@@ -17,7 +27,7 @@ class FSiRemovableFileTypeSpec extends ObjectBehavior
 
     function it_is_resource()
     {
-        $this->shouldImplement('FSi\Bundle\ResourceRepositoryBundle\Repository\Resource\Type\ResourceInterface');
+        $this->shouldImplement(ResourceInterface::class);
     }
 
     function it_return_text_entity_field()
@@ -32,10 +42,10 @@ class FSiRemovableFileTypeSpec extends ObjectBehavior
 
     function it_return_form_builder(FormFactory $factory, FormBuilder $form)
     {
-        $factory->createNamedBuilder('fileValue', 'fsi_removable_file', null, array(
+        $factory->createNamedBuilder('fileValue', RemovableFileType::class, null, [
             'label' => false,
             'required' => false,
-        ))->shouldBeCalled()->willReturn($form);
+        ])->shouldBeCalled()->willReturn($form);
 
         $this->getFormBuilder($factory)->shouldReturn($form);
     }
@@ -44,30 +54,30 @@ class FSiRemovableFileTypeSpec extends ObjectBehavior
     {
         $this->addConstraint($constraint);
 
-        $factory->createNamedBuilder('fileValue', 'fsi_removable_file', null, array(
+        $factory->createNamedBuilder('fileValue', RemovableFileType::class, null, [
             'label' => false,
             'required' => false,
-            'file_options' => array(
-                'constraints' => array(
+            'file_options' => [
+                'constraints' => [
                     $constraint
-                )
-            )
-        ))->shouldBeCalled()->willReturn($form);
+                ]
+            ]
+        ])->shouldBeCalled()->willReturn($form);
 
         $this->getFormBuilder($factory)->shouldReturn($form);
     }
 
     function it_return_form_builder_with_form_options_added_to_resource_definition(FormFactory $factory, FormBuilder $form)
     {
-        $this->setFormOptions(array(
+        $this->setFormOptions([
             'file_type' => 'fsi_image'
-        ));
+        ]);
 
-        $factory->createNamedBuilder('fileValue', 'fsi_removable_file', null, array(
+        $factory->createNamedBuilder('fileValue', RemovableFileType::class, null, [
             'label' => false,
             'required' => false,
             'file_type' => 'fsi_image'
-        ))->shouldBeCalled()->willReturn($form);
+        ])->shouldBeCalled()->willReturn($form);
 
         $this->getFormBuilder($factory)->shouldReturn($form);
     }
@@ -79,20 +89,20 @@ class FSiRemovableFileTypeSpec extends ObjectBehavior
     ) {
         $this->addConstraint($constraint);
 
-        $this->setFormOptions(array(
+        $this->setFormOptions([
             'file_type' => 'fsi_image',
-            'file_options' => array('file_option' => 'value')
-        ));
+            'file_options' => ['file_option' => 'value']
+        ]);
 
-        $factory->createNamedBuilder('fileValue', 'fsi_removable_file', null, array(
+        $factory->createNamedBuilder('fileValue', RemovableFileType::class, null, [
             'label' => false,
             'required' => false,
             'file_type' => 'fsi_image',
-            'file_options' => array(
+            'file_options' => [
                 'file_option' => 'value',
-                'constraints' => array($constraint)
-            ),
-        ))->shouldBeCalled()->willReturn($form);
+                'constraints' => [$constraint]
+            ],
+        ])->shouldBeCalled()->willReturn($form);
 
         $this->getFormBuilder($factory)->shouldReturn($form);
     }
